+++ projects/angular-sdk/src/i18n/en.ts
@@ -262,7 +262,7 @@ export const locale: I18n = {
                 CLOSE: 'Close'
             },
             TERMS_LINK_TEXT: "Terms & Conditions apply",
-            COVERAGE_PERCENTAGE: "{{ coverage }} Flight cost covered",
+            COVERAGE_PERCENTAGE: "{{ coverage }} flight cost covered",
             PRICE_PER_TRAVELER: "<span class='price-per-traveler'>{{ price }}</span><span>per traveler</span>"
         },
         CFAR_OFFER_BANNER: {
@@ -290,10 +290,10 @@ export const locale: I18n = {
             TERMS_WARNING: "If you don't add this, your ticket is non-refundable."
         },
         CFAR_EXERCISE_FLOW: {
-            TITLE: "Cancel flight",
+            TITLE: "Cancel your flight with Hopper",
             SUBTITLE: "Cancel For Any Reason",
-            DEADLINE_SENTENCE_1: "Cancel your booking free of charge until ",
-            DEADLINE_SENTENCE_2: "You'll receive a confirmation email and the trip status will be updated.",
+            DEADLINE_SENTENCE_1: "You have purchased a Cancel For Any Reason policy with your booking which enables you to cancel your booking free of charge until ",
+            DEADLINE_SENTENCE_2: "Don't forget that the cancellation is only possible for the entire trip and all the passengers in the booking. Based on your coverage amount, you will be refunded of a total of {{ amount }}.",
             TO: " to ",
             DEPARTURE: "Departure:",
             ARRIVAL: "Arrival:",
@@ -301,14 +301,14 @@ export const locale: I18n = {
             PASSENGER: "{{ nbPassenger }} passenger",
             PASSENGERS: "{{ nbPassenger }} passengers",
             REFUND_DETAILS_TITLE: "Refund details",
-            REFUND_WARNING_MESSAGE: "You have purchased a Cancel For Any Reason policy with your booking, which enables you to cancel and get refunded of {{ amount }}. If you want to cancel your booking for the entire trip and all the passengers, you can start the refund process below",
+            REFUND_WARNING_MESSAGE: "If you are sure you want to proceed to the cancellation and get your refund, please continue.",
             HOPPER_REFUND: "Hopper refund",
             AIRLINE_REFUND: "Airline refund",
             METHOD: "Method:",
             AMOUNT: "Airline:",
             REFUND_METHOD: "Refund method",
             CHECK_VERIFICATION_TITLE: "Identity verification",
-            CHECK_VERIFICATION_TEXT: "In order to verify your identity, we will send a code to your email address",
+            CHECK_VERIFICATION_TEXT: "In order to verify your identity, we will send a code to your email address (the one used when booking your flight).",
             SEND_VERIFICATION_TITLE: "Identity verification",
             SEND_VERIFICATION_TEXT: "We just sent an email with a verification code to {{email}}. Please enter that code below",
             BUTTON: {
@@ -322,8 +322,8 @@ export const locale: I18n = {
             },
             NO_CONTRACT: "No contract available",
             LOADING: "Loading in progress...",
-            LOADING_HYPERWALLET: "Loading Hyperwallet in progress...",
-            LOADING_HYPERWALLET_ERROR: "An error occured while loading Hyperwallet",
+            LOADING_HYPERWALLET: "Loading in progress...",
+            LOADING_HYPERWALLET_ERROR: "An error occured, please contact our customer support via <a href='{{ link }}' target='_blank'>this form.</a>",
             STEP: {
                 FIRST: "Booking summary",
                 SECOND: "General information",
@@ -355,7 +355,7 @@ export const locale: I18n = {
                 ZIP_ERROR: "Zip/Postal Code is required"
             },
             CONFIRMATION_MESSAGE_1: "Your flight has been canceled and your refund is being processed.",
-            CONFIRMATION_MESSAGE_2: "A confirmation email will be sent shortly to {{ userEmail }} with details on how to use your airlines credits",
+            CONFIRMATION_MESSAGE_2: "A confirmation email will be sent shortly to {{ userEmail }}",
             REFUND_DETAILS: "{{ amount }} in {{ method }}",
             PASSENGER_TYPE: {
                 ADULT: "Adult",