+++ projects/angular-sdk/src/i18n/es.ts
@@ -261,22 +261,21 @@ export const locale: I18n = {
             BUTTON: {
                 CLOSE: 'Cerrar'
             },
-            TERMS_LINK_TEXT: "Términos y Condiciones"
+            TERMS_LINK_TEXT: "Términos y Condiciones",
+            COVERAGE_PERCENTAGE: "{{ coverage }} de los costos de vuelo cubiertos",
+            PRICE_PER_TRAVELER: "<span class='price-per-traveler'>{{ price }}</span><span>por pasajero</span>",
         },
         CFAR_OFFER_BANNER: {
             TITLE: "Cancele su vuelo en cualquier momento, por cualquier motivo.",
             LOADING: "Cargando ofertas en curso...",
             BUTTON: {
                 SELECT: "Seleccione"
-            },
-            COVERAGE_PERCENTAGE: "{{ coverage }} de los costos de vuelo cubiertos",
+            }
         },
         CFAR_OFFER_BANNER_LARGE: {
             TITLE: "Cancele su vuelo en cualquier momento, por cualquier motivo.",
             LOADING: "Cargando ofertas en curso...",
-            COVERAGE_PERCENTAGE: "{{ coverage }} de los costos de vuelo cubiertos",
             NO_COVERAGE_OPTION: "No quiero comprar Cancelar por cualquier motivo.",
-            PRICE_PER_TRAVELER: "<span class='price-per-traveler'>{{ price }}</span><span>por pasajero</span>",
             WARNING_MESSAGE: "Elija o rechace Cancelar por cualquier motivo."
         },
         CFAR_OFFER_DIALOG: {
@@ -290,7 +289,7 @@ export const locale: I18n = {
             LOADING: "Cargando ofertas en curso...",
             TERMS_WARNING: "Si no agrega esto, su boleto no es reembolsable"
         },
-        CFAR_EXERCISE_DIALOG: {
+        CFAR_EXERCISE_FLOW: {
             TITLE: "Cancelar vuelo",
             SUBTITLE: "Cancelar por cualquier motivo",
             DEADLINE_SENTENCE_1: "Cancela tu reserva sin cargo hasta ",
@@ -363,7 +362,8 @@ export const locale: I18n = {
                 CHILD: "Niño",
                 SEATED_INFANT: "Bebé sentado",
                 LAP_INFANT: "Infante de regazo"
-            }
+            },
+            CONTRACT_ALREADY_REFUNDED: "Tu solicitud de cancelación ya ha sido tenida en cuenta y su reembolso está en curso. Si necesita ayuda, puede ponerse en contacto con nuestro equipo de atención al cliente a través de <a href='{{ link }}' target='_blank'>este formulario.</a>"
         }
     }
 }