+++ projects/angular-sdk/CHANGELOG.md
@@ -3,8 +3,11 @@
 # 2.8.2 (2022-XX-XX)
 
 ### Bug fixes / Small changes
-- Offer Banner Large: 
+- CFAR Offer Banner Large: 
     - Fixed the display of warning message
+    - Added a message when the booking has already been refunded
+- CFAR Offer Dialog:
+    - Fixed the display of price per traveler
 - General:
     - Fixed XSS vulnerability